+++ cmd/thrust/main.go
@@ -13,8 +13,6 @@ func setLogger(level slog.Level) {
 }
 
 func main() {
-	setLogger(slog.LevelWarn)
-
 	config, err := internal.NewConfig()
 	if err != nil {
 		fmt.Printf("ERROR: %s\n", err)