+++ internal/config.go
@@ -61,12 +61,6 @@ type Config struct {
 	LogLevel slog.Level
 }
 
-var warn func(msg string, args ...any)
-
-func init() {
-	warn = slog.Warn
-}
-
 func NewConfig() (*Config, error) {
 	if len(os.Args) < 2 {
 		return nil, errors.New("missing upstream command")
@@ -76,12 +70,6 @@ func NewConfig() (*Config, error) {
 	if getEnvBool("DEBUG", false) {
 		logLevel = slog.LevelDebug
 	}
-
-	sslDomain := getEnvString("SSL_DOMAIN", "")
-	if sslDomain != "" {
-		warn("SSL_DOMAIN is deprecated. Use TLS_DOMAIN instead.", "SSL_DOMAIN", sslDomain)
-	}
-	tlsDomain := getEnvString("TLS_DOMAIN", sslDomain)
 
 	return &Config{
 		TargetPort:      getEnvInt("TARGET_PORT", defaultTargetPort),
@@ -93,7 +81,7 @@ func NewConfig() (*Config, error) {
 		XSendfileEnabled:      getEnvBool("X_SENDFILE_ENABLED", true),
 		MaxRequestBody:        getEnvInt("MAX_REQUEST_BODY", defaultMaxRequestBody),
 
-		TLSDomain:        tlsDomain,
+		TLSDomain:        getEnvString("TLS_DOMAIN", ""),
 		ACMEDirectoryURL: getEnvString("ACME_DIRECTORY", defaultACMEDirectoryURL),
 		EAB_KID:          getEnvString("EAB_KID", ""),
 		EAB_HMACKey:      getEnvString("EAB_HMAC_KEY", ""),